Update inventory handling and refactor classes
Significantly refactored the handling of inventories and tabs in order to make the code more manageable. This included renaming 'Formatting.cs' to 'LocalizedTextManager.cs' to better reflect its functionality, as well as simplifying the way that items are displayed and selected within the inventory UI. Stripped out unnecessary sections of code, and updated old sections to fit with the new structure. The code now uses more dependable ways to call on the current inventory and tabs. Furthermore, it disabled background fade for the active menu and redrew foregrounds to enhance visibility.