Update storage information and menu handling
This update includes changes to improve storage information for different types such as building, location, and big craftable storage. Code was added to provide specific names and descriptions for each storage type. Furthermore, changes were made to the handling of menus, specifically in the ItemGrabMenuManager. The aim was to enhance the update and render processes for menus. Some minor adjustments were also made to Keybind configuration in the Controls class, and a minor event handling error was fixed in the ChestInfo feature.