Remove DisplayedItems and add Pointer classes in BetterChests
The DisplayedItems.cs file was removed as part of a larger refactor for improving the UI. A new Pointer.cs file was introduced to provide a cleaner way of pointing to objects on the map. Additionally, changes were made to the ChestFinder.cs file to replace the usage of ChestContainers with Pointers.