Update method execution and keypress handling logic
The primary changes in this commit involve adjusting and reordering the way features are interacted with by means of textbox options in ConfigMenu, as well as altering the way keypresses are handled. The update also corrects a misnamed method and enhances the way the application counts item stacks to check the capacity of destination containers. The structure of some features has also undergone modifications.