Refactor mod configuration management and minor improvements
The commit introduces a significant refactor to the way mod configurations are managed. GenericModConfigMenu handling has been improved and the ConfigManager class is now more intuitive. There are also minor changes to ProxyChestFactory and ProxyChestRequest classes, such as minor type adjustments for better compatibility. As part of the refactor, the ConfigChangedEventArgs class has been moved from the BetterChests project to the Common project for better reuse.