Refactor codebase and fix bugs
Made several changes to improve code functionality, maintainability, and readability. Removed redundant codes and files like the Logging.cs, fixed bugs, and renamed variables to adhere to standards. Also implemented a new feature for taking all tiles in a box centered on the player character for garbage collection and dispensing services in the EasyAccess/Framework.