Add support for complex menu options and enhance storage features
The main change in this commit is the addition of support for complex menu options through the implementation of IComplexOption interface. Moreover, the temporary storage options were updated to be able to save changes back to the storage options and reset them to their defaults. Optimizations were also made to the "CategorizeChest" feature as well as some updates to the "LabelChest" functionality. Additionally, bugs and irregularities were addressed in various parts of the system.