Update version and modify ConfigManager registration
The version in the FuryCore project file has been updated to the latest date. In the ConfigManager class, the method call to register a mod's configuration has been modified to exclude the manifest parameter. The changes are likely meant to reconcile the existing code with updates in dependencies or conventions.