Update method argument descriptions and improve mod registration
The descriptions of modRegistry arguments in multiple classes have been updated for more clarity. The GenericModConfigMenuIntegration's registration methods are now more secure, only allowing each mod to be registered/unregistered once. Furthermore, the position of some file references in Common.projitems has been adjusted, and the IGenericModConfigMenuApi.cs file has been modified to prevent the addition of an unnecessary newline at the end of the file.