Update mods for final SDV 1.6 and SMAPI 4.0.0
The modifications in multiple mods have been carried out to be compatible with the final version of Stardew Valley 1.6 and SMAPI 4.0.0. Some unused files have been removed, and some projects and event handling are also refactored to enhance functionality. Several minor bug fixes and updates in changelogs and mods properties are included as well.